상용서버 반영

+++ app-admin-api/src/main/java/co/kurrant/app/admin_api/dto/schedules/ItemPageableResponseDto.java
@@ -18,7 +18,7 @@ public class ItemPageableResponseDto<T> {
 
     @Min(0)
     @NotNull
-    private Long offset;
+    private Integer offset;
 
     @Min(0)
     @NotNull
